Add logging functionality to BookService and PageService
This commit introduces a private logger of type `ILogger<T>` to both the `BookService` and `PageService` classes. The constructors have been updated to accept the logger as a parameter, enabling logging capabilities within the services. Additionally, the `using` directive for `Microsoft.Extensions.Logging` has been added to both files to support this functionality.